requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
mysql-connector-python==8.3.0
selenium==4.16.0
scikit-learn==1.4.0
xgboost==2.0.3
catboost==1.2.2
transformers==4.37.0
torch==2.1.2
